+++ src/textual/widgets/_tabbed_content.py
@@ -42,6 +42,7 @@ class TabPane(Widget):
         height: auto;
         padding: 1 2;
     }
+
     """
 
     def __init__(